Provide Judoscale functionality as a Flask extension
