+++ judoscale/flask/__init__.py
@@ -1,0 +1,3 @@
+from judoscale.flask.judoscale import Judoscale
+
+__all__ = ["Judoscale"]
