+++ sample-apps/flask_sample/app.py
@@ -1,28 +1,24 @@
-import logging
-from judoscale.core.config import config as judoconfig
-from judoscale.flask.middleware import RequestQueueTimeMiddleware
-
-from flask import Flask, request
 import settings
+from flask import Flask, current_app
 
-logger = logging.getLogger(__name__)
+from judoscale.flask import Judoscale
 
-
-app = Flask("DemoFlaskApp")
-config_obj = settings.BaseConfig
-app.config.from_object(config_obj)
-
-judoconfig.merge(getattr(config_obj, "JUDOSCALE", {}))
-
-# calling our middleware
-app.wsgi_app = RequestQueueTimeMiddleware(app.wsgi_app)
+judoscale = Judoscale()
 
 
-@app.route("/", methods=["GET"])
-def index():
-    logger.warning("Hello, world")
-    return "Judoscale Flask Sample App"
+def create_app():
+    app = Flask("DemoFlaskApp")
+    app.config.from_object(settings.BaseConfig)
+
+    judoscale.init_app(app)
+
+    @app.route("/", methods=["GET"])
+    def index():
+        current_app.logger.warning("Hello, world")
+        return "Judoscale Flask Sample App"
+
+    return app
 
 
 if __name__ == "__main__":
-    app.run("127.0.0.1", "5000", debug=True)
+    create_app().run("127.0.0.1", "5000", debug=True)
